Add DAI to list of cryptos to have buyCryptoModal
