+++ src/constants/WalletAndCurrencyConstants.js
@@ -1,5 +1,5 @@
 // @flow
-
+import s from '../locales/strings.js'
 export const ETHEREUM_WALLET = 'wallet:ethereum'
 export const BITCOIN_WALLET = 'wallet:bitcoin'
 export const BITCOINCASH_WALLET = 'wallet:bitcoincash'
@@ -31,6 +31,101 @@ export const CURRENCY_PLUGIN_NAMES = {
   UFO: 'ufo',
   VTC: 'vertcoin',
   XZC: 'zcoin'
+}
+
+export const getSpecialCurrencyInfo = (currencyCode: string): Object => {
+  if (SPECIAL_CURRENCY_INFO[currencyCode]) {
+    return SPECIAL_CURRENCY_INFO[currencyCode]
+  } else {
+    return {}
+  }
+}
+
+type SpecialCurrencyInfo = {
+  [currencyCode: string]: {
+    noMaxSpend?: boolean,
+    needsAccountNameSetup?: boolean,
+    noChangeMiningFee?: boolean,
+    allowZeroTx?: boolean,
+    dummyPublicAddress?: string,
+    uniqueIdentifier?: {
+      addButtonText: string,
+      identifierName: string,
+      identifierKeyboardType: string
+    },
+    minimumPopupModals?: {
+      minimumNativeBalance: string,
+      modalMessage: string
+    }
+  }
+}
+
+export const SPECIAL_CURRENCY_INFO: SpecialCurrencyInfo = {
+  BTC: {
+    displayBuyCrypto: true
+  },
+  BCH: {
+    displayBuyCrypto: true
+  },
+  LTC: {
+    displayBuyCrypto: true
+  },
+  XLM: {
+    dummyPublicAddress: 'GBEVGJYAUKJ2TVPMC3GEPI2GGZQLMWZDRWJCVNBXCJ3ELYTDPHVQQM74',
+    noCustomMiningFee: true,
+    uniqueIdentifier: {
+      addButtonText: s.strings.unique_identifier_dropdown_option_memo_id,
+      identifierName: s.strings.unique_identifier_memo_id,
+      identifierKeyboardType: 'numeric'
+    },
+    minimumPopupModals: {
+      minimumNativeBalance: '10000000',
+      modalMessage: s.strings.request_xlm_minimum_notification_body
+    },
+    displayBuyCrypto: false
+  },
+  XRP: {
+    dummyPublicAddress: 'rfuESo7eHUnvebxgaFjfYxfwXhM2uBPAj3',
+    noCustomMiningFee: true,
+    uniqueIdentifier: {
+      addButtonText: s.strings.unique_identifier_dropdown_option_destination_tag,
+      identifierName: s.strings.unique_identifier_destination_tag,
+      identifierKeyboardType: 'numeric'
+    },
+    minimumPopupModals: {
+      minimumNativeBalance: '20000000',
+      modalMessage: s.strings.request_xrp_minimum_notification_body
+    },
+    displayBuyCrypto: true
+  },
+  XMR: {
+    dummyPublicAddress: '46qxvuS78CNBoiiKmDjvjd5pMAZrTBbDNNHDoP52jKj9j5mk6m4R5nU6BDrWQURiWV9a2n5Sy8Qo4aJskKa92FX1GpZFiYA',
+    noCustomMiningFee: true,
+    noMaxSpend: true,
+    uniqueIdentifier: {
+      addButtonText: s.strings.unique_identifier_dropdown_option_payment_id,
+      identifierName: s.strings.unique_identifier_payment_id,
+      identifierKeyboardType: 'default'
+    }
+  },
+  EOS: {
+    dummyPublicAddress: 'edgecreator2',
+    needsAccountNameSetup: true,
+    noChangeMiningFee: true,
+    uniqueIdentifier: {
+      addButtonText: s.strings.unique_identifier_dropdown_option_memo,
+      identifierName: s.strings.unique_identifier_memo,
+      identifierKeyboardType: 'default'
+    }
+  },
+  ETH: {
+    dummyPublicAddress: '0x0d73358506663d484945ba85d0cd435ad610b0a0',
+    allowZeroTx: true,
+    displayBuyCrypto: true
+  },
+  DAI: {
+    displayBuyCrypto: true
+  }
 }
 
 export const USD_FIAT = 'iso:USD'
